docker/predictor/requirements.txt
darkskylib
asyncio
keras
pandas
numpy
matplotlib
sklearn
influxdb
tensorflow